requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.24
psycopg2-binary==2.9.9